commands/service:compile: update the compilation structure
* moved definition types to service-command.ts file.
